updated review points.
