--- Engine/src/main/java/com/ing/engine/totp/EnvLoader.java
@@ -1,27 +1,0 @@
-package com.ing.engine.totp;
-
-
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-
-public class EnvLoader {
-
-    public static Map<String, String> loadEnv(String filePath) throws IOException {
-        Map<String, String> env = new HashMap<>();
-        try (BufferedReader reader = new BufferedReader(new FileReader(filePath))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                if (!line.trim().startsWith("#") && line.contains("=")) {
-                    String[] parts = line.split("=", 2);
-                    env.put(parts[0].trim(), parts[1].trim());
-                }
-            }
-        }
-        return env;
-    }
-
-}
